+++ python_modules/libraries/dagster-aws/dagster_aws/pipes/__init__.py
@@ -1,0 +1,20 @@
+from dagster_aws.pipes.clients import PipesGlueClient, PipesLambdaClient
+from dagster_aws.pipes.context_injectors import (
+    PipesLambdaEventContextInjector,
+    PipesS3ContextInjector,
+)
+from dagster_aws.pipes.message_readers import (
+    PipesCloudWatchMessageReader,
+    PipesLambdaLogsMessageReader,
+    PipesS3MessageReader,
+)
+
+__all__ = [
+    "PipesGlueClient",
+    "PipesLambdaClient",
+    "PipesS3ContextInjector",
+    "PipesLambdaEventContextInjector",
+    "PipesS3MessageReader",
+    "PipesLambdaLogsMessageReader",
+    "PipesCloudWatchMessageReader",
+]
