+++ python_modules/libraries/dagster-aws/dagster_aws/pipes/clients/__init__.py
@@ -1,0 +1,4 @@
+from dagster_aws.pipes.clients.glue import PipesGlueClient
+from dagster_aws.pipes.clients.lambda_ import PipesLambdaClient
+
+__all__ = ["PipesGlueClient", "PipesLambdaClient"]
